requirements.txt
cudatoolkit=10.2
pytorch==1.5.1
transformers==2.8.0
